// packages/contracts/src/generated/exchange.ts
export const AddressesRow = z.object({
  "id": z.string().uuid(),
  "user_id": z.string().uuid().nullable(),
  "line_1": z.string().nullable(),
  "line_2": z.string().nullable(),
  "city": z.string().nullable(),
  "state": z.string().nullable(),
  "country": z.string().nullable(),
  "zip": z.string().nullable(),
  "created_at": z.string().nullable(),
  "updated_at": z.string().nullable(),
  "name": z.string().nullable(),
  "is_default": z.boolean().nullable(),
  "phone_number": z.string().nullable(),
  "is_valid": z.boolean().nullable(),
  "country_code": z.string().nullable(),
  "is_residential": z.boolean().nullable(),
});
export type AddressesRow = z.infer<typeof AddressesRow>;